chore: move non_constructible.h to google/cloud/internal
This type will be generally useful when converting interfaces to
use `google::cloud::Options`.
