+++ google/cloud/pubsub/CMakeLists.txt
@@ -55,7 +55,6 @@ add_library(
     internal/defaults.h
     internal/flow_controlled_publisher_connection.cc
     internal/flow_controlled_publisher_connection.h
-    internal/non_constructible.h
     internal/ordering_key_publisher_connection.cc
     internal/ordering_key_publisher_connection.h
     internal/publisher_auth.cc
