+++ google/cloud/pubsub/subscription_admin_connection.h
@@ -17,12 +17,12 @@
 
 #include "google/cloud/pubsub/backoff_policy.h"
 #include "google/cloud/pubsub/connection_options.h"
-#include "google/cloud/pubsub/internal/non_constructible.h"
 #include "google/cloud/pubsub/internal/subscriber_stub.h"
 #include "google/cloud/pubsub/retry_policy.h"
 #include "google/cloud/pubsub/snapshot.h"
 #include "google/cloud/pubsub/subscription.h"
 #include "google/cloud/pubsub/version.h"
+#include "google/cloud/internal/non_constructible.h"
 #include "google/cloud/internal/pagination_range.h"
 #include "google/cloud/status_or.h"
 #include <google/pubsub/v1/pubsub.pb.h>
@@ -196,7 +196,7 @@ class SubscriptionAdminConnection {
  * @deprecated Please use `MakeSubscriptionAdminConnection()` instead.
  */
 std::shared_ptr<SubscriptionAdminConnection> MakeSubscriptionAdminConnection(
-    std::initializer_list<pubsub_internal::NonConstructible>);
+    std::initializer_list<internal::NonConstructible>);
 
 /**
  * Creates a new `SubscriptionAdminConnection` object to work with
